+++ chain/chain/src/resharding/mod.rs
@@ -2,5 +2,5 @@ pub mod event_type;
 pub mod flat_storage_resharder;
 pub mod manager;
 pub mod resharding_actor;
-pub mod state_resharder;
+pub mod trie_state_resharder;
 pub mod types;
